verifysense/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-api-python-client==2.97.0
google-auth==2.22.0
google-cloud-vision==3.4.3
google-cloud-firestore==2.11.1
google-generativeai==0.3.1